Update Card.java
I added the javadoc back for rank and suit. I'm not sure if this is correct, I saw they had an @link in the comments. I got rid of methods to ensure @NotNull as the unit test expects IllegalArgumentException to be thrown. Got rid of excess asterisks in the javadocs. I'm not sure if there is supposed to be a javadoc for variables as well?